Download Weather Alert Table from Senamhi

This function downloads the table of weather warnings provided by Senamhi For more information, please visit the following link: https://www.Senamhi.gob.pe/?&p=aviso-meteorologico

Usage

extract_meteorological_table()

Value

A tibble object containing the weather alert data.

Examples

# \donttest{
library(geoidep)
extract_meteorological_table() |> head()
#> Reading data from the website...
#> Data successfully uploaded.
#> # A tibble: 6 × 7
#>   aviso                                nro   emision inicio fin   duracion nivel
#>   <chr>                                <chr> <chr>   <chr>  <chr> <chr>    <chr>
#> 1 PRECIPITACIONES EN LA SIERRA         295 … 2024-1… 2024-… 2024… 47 Hrs.  NARA…
#> 2 DESCENSO DE TEMPERATURA NOCTURNA EN… 294 … 2024-1… 2024-… 2024… 71 Hrs.  ROJO 
#> 3 INCREMENTO DE VIENTO EN LA COSTA     293 … 2024-1… 2024-… 2024… 59 Hrs.  NARA…
#> 4 PRECIPITACIONES EN LA SIERRA CENTRO… 292 … 2024-1… 2024-… 2024… 23 Hrs.  NARA…
#> 5 INCREMENTO DE VIENTO EN LA SIERRA    291   2024-1… 2024-… 2024… 59 Hrs.  NARA…
#> 6 NEVADA EN LA SIERRA CENTRO Y SUR     290   2024-1… 2024-… 2024… 47 Hrs.  AMAR…
# }