Download the Departament limits of INEI

This function allows you to download the latest version of the geometry and ubigeos corresponding to the official political division of the departament boundaries of Peru. For more information, you can visit the following page https://ide.inei.gob.pe/#capas

Usage

get_departaments(dsn = NULL, show_progress = TRUE, quiet = TRUE)

Arguments

dsn

Character. Output filename. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

A sf or tibble object.

Examples

library(geoidep)
dist <- get_departaments(show_progress = FALSE)
#> Error in curl::curl_fetch_disk(url, x$path, handle = handle): SSL peer certificate or SSH remote key was not OK: [ide.inei.gob.pe] SSL certificate problem: unable to get local issuer certificate
head(dist)
#>                                                                     
#> 1 function (x, method = "euclidean", diag = FALSE, upper = FALSE,   
#> 2     p = 2)                                                        
#> 3 {                                                                 
#> 4     if (!is.na(pmatch(method, "euclidian")))                      
#> 5         method <- "euclidean"                                     
#> 6     METHODS <- c("euclidean", "maximum", "manhattan", "canberra",