Download information on the latest deforestation alerts detected by Geobosque

This function allows you to download deforestation alert information detected by Geobosque for any polygon located in Peru. For more details, please visit the following website: https://geobosques.minam.gob.pe

Usage

get_early_warning(region, sf = TRUE, show_progress = TRUE)

Arguments

region

A string Specifies the unique geographical code of interest.

sf

Logical. Indicates whether to return the data as an `sf` object.

show_progress

Logical. If TRUE, shows a progress bar during download.

Value

A tibble or sf object, depending on the value of `sf`.

Examples

# \donttest{
library(geoidep)
amazonas <- get_departaments(show_progress = FALSE) |> subset(NOMBDEP == "AMAZONAS")
warning_point <- get_early_warning(region = amazonas, sf = TRUE, show_progress = FALSE)
head(warning_point)
#> Simple feature collection with 6 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -77.92989 ymin: -3.07257 xmax: -77.82786 ymax: -3.011245
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 2
#>   descrip                     geometry
#>   <chr>                    <POINT [°]>
#> 1 Warning points (-77.83573 -3.011245)
#> 2 Warning points (-77.83329 -3.034038)
#> 3 Warning points (-77.88837 -3.040279)
#> 4 Warning points (-77.82786 -3.048962)
#> 5 Warning points (-77.92989 -3.072299)
#> 6 Warning points  (-77.92908 -3.07257)
# }