Download available forest fire data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_forest_fire_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
fire_forest <- get_forest_fire_data(show_progress = FALSE)
#> Reading layer `file1a831d01ecd6' from data source 
#>   `/tmp/Rtmpwh3zuB/file1a831d01ecd6.geojson' using driver `GeoJSON'
#> Simple feature collection with 386 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -79.10685 ymin: -15.43997 xmax: -69.93638 ymax: -4.03436
#> Geodetic CRS:  WGS 84
head(fire_forest)
#> Simple feature collection with 6 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -77.00076 ymin: -14.86937 xmax: -71.53812 ymax: -8.97724
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1   211126 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-06
#> 2   211253 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-07
#> 3   211372 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-07
#> 4   211468 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-07
#> 5   211127 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-06
#> 6   211128 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-06
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN  NOMDEP      NOMPRO    NOMDIS   CAPITAL      FECHA     HORA
#> 1     19      2   CUSCO     ESPINAR COPORAQUE COPORAQUE 2024-10-06 13:11:00
#> 2     19      2   CUSCO     ACOMAYO      ACOS      ACOS 2024-10-06 22:08:00
#> 3     19      2   CUSCO     ACOMAYO      ACOS      ACOS 2024-10-07 13:15:00
#> 4     18      2 HUANUCO HUACAYBAMBA     PINRA     PINRA 2024-10-07 14:10:00
#> 5     19      2   CUSCO     ESPINAR COPORAQUE COPORAQUE 2024-10-06 13:11:00
#> 6     19      2   CUSCO     ESPINAR COPORAQUE COPORAQUE 2024-10-06 13:11:00
#>   CATEG           NOMCATEG   LATITUD  LONGITUD   COORES  COORNO         CODREP
#> 1    25           APACHACO -14.86937 -71.54598 226043.9 8354559 PIF_04443_2024
#> 2    25 CORMA LLANSA CORMA -13.90889 -71.75365 202412.4 8460630 PIF_04445_2024
#> 3    25 CORMA LLANSA CORMA -13.91017 -71.74924 202890.9 8460494 PIF_04445_2024
#> 4    30                     -8.97724 -77.00076 280031.9 9007065 PIF_04480_2024
#> 5    25           APACHACO -14.86881 -71.54205 226466.3 8354626 PIF_04443_2024
#> 6    25           APACHACO -14.86825 -71.53812 226888.7 8354693 PIF_04443_2024
#>   created_user        created_date last_edited_user    last_edited_date ESTADO
#> 1   JMONTALBAN 2024-10-07 06:04:10           UFMS17 2024-10-08 00:43:14      2
#> 2   JMONTALBAN 2024-10-07 14:18:02           UFMS17 2024-10-08 01:12:58      2
#> 3   JMONTALBAN 2024-10-08 00:25:11           UFMS17 2024-10-08 01:12:58      2
#> 4   JMONTALBAN 2024-10-08 00:25:11       JMONTALBAN 2024-10-08 00:25:11      1
#> 5   JMONTALBAN 2024-10-07 06:04:10           UFMS17 2024-10-08 00:43:14      2
#> 6   JMONTALBAN 2024-10-07 06:04:10           UFMS17 2024-10-08 00:43:14      2
#>               TIPCOB PELIGRO REFESTADO CONF CODRVIF SENSAT CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal       0         0    0       0      N     08   0808 080803
#> 2 Cobertura Forestal       0         0    0       0      T     08   0802 080203
#> 3 Cobertura Forestal       0         0    0       0    N20     08   0802 080203
#> 4 Cobertura Forestal       0         0    0       0    N21     10   1004 100404
#> 5 Cobertura Forestal       0         0    0       0      N     08   0808 080803
#> 6 Cobertura Forestal       0         0    0       0      N     08   0808 080803
#>                      geometry
#> 1 POINT (-71.54598 -14.86937)
#> 2 POINT (-71.75365 -13.90889)
#> 3 POINT (-71.74924 -13.91017)
#> 4  POINT (-77.00076 -8.97724)
#> 5 POINT (-71.54205 -14.86881)
#> 6 POINT (-71.53812 -14.86825)
# }