Download the geographic information from MIDAGRI

This function downloads the latest version of the geographic data representing the agricultural area of Peru. For more details, please visit the following page: https://siea.midagri.gob.pe/portal/informativos/superficie-agricola-peruana.

Usage

get_midagri_data(dsn = NULL, layer = NULL, show_progress = TRUE, quiet = TRUE)

Arguments

dsn

Character. The output filename in *.gpkg format. If not provided, a temporary file will be created.

layer

Character. Currently, only one layer is available.

show_progress

Logical. If TRUE, displays a progress bar.

quiet

Logical. If TRUE, suppresses informational messages.

Value

An `sf` object containing the downloaded geographic data.

Details

Available layer:

vegetation cover: The agricultural area of Peru in shapefile format, provided at the national level by SIEA.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
# Disable the use of S2 geometry for spatial operations
sf_use_s2(use_s2 = FALSE)
#> Spherical geometry (s2) switched off

# Downloading data and filtering by department
junin <- get_departaments(show_progress = FALSE) |>
  subset(NOMBDEP == "JUNIN")

# Performing an intersection between the Junin department and vegetation cover
cov_veg <- get_midagri_data(layer = "vegetation cover", show_progress = FALSE) |>
  st_intersection(junin)
#> Error in curl::curl_fetch_disk(url, x$path, handle = handle): Timeout was reached: [archivos.midagri.gob.pe] Connection timeout after 10000 ms

# Plotting the geometry of the intersected vegetation cover
plot(st_geometry(cov_veg))
#> Error in eval(expr, envir, enclos): object 'cov_veg' not found
# }