Download available forest fire data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_forest_fire_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
fire_forest <- get_forest_fire_data(show_progress = FALSE)
#> Reading layer `file1ab7501da463' from data source 
#>   `/tmp/Rtmpxr6CpH/file1ab7501da463.geojson' using driver `GeoJSON'
#> Simple feature collection with 392 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -79.49467 ymin: -16.34176 xmax: -69.17239 ymax: -4.79419
#> Geodetic CRS:  WGS 84
head(fire_forest)
#> Simple feature collection with 6 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -74.14829 ymin: -8.60201 xmax: -74.13892 ymax: -8.60107
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1   214053 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 2   214054 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 3   214055 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 4   214056 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 5   214057 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 6   214058 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN  NOMDEP           NOMPRO  NOMDIS CAPITAL      FECHA     HORA
#> 1     18      2 UCAYALI CORONEL PORTILLO MASISEA MASISEA 2024-10-18 00:36:00
#> 2     18      2 UCAYALI CORONEL PORTILLO MASISEA MASISEA 2024-10-18 01:27:00
#> 3     18      2 UCAYALI CORONEL PORTILLO MASISEA MASISEA 2024-10-17 00:53:00
#> 4     18      2 UCAYALI CORONEL PORTILLO MASISEA MASISEA 2024-10-17 01:46:00
#> 5     18      2 UCAYALI CORONEL PORTILLO MASISEA MASISEA 2024-10-18 01:27:00
#> 6     18      2 UCAYALI CORONEL PORTILLO MASISEA MASISEA 2024-10-17 00:53:00
#>   CATEG NOMCATEG  LATITUD  LONGITUD   COORES  COORNO         CODREP
#> 1    11   Imiria -8.60201 -74.14829 593718.2 9049044 PIF_04668_2024
#> 2    11   Imiria -8.60177 -74.14085 594537.0 9049069 PIF_04668_2024
#> 3    11   Imiria -8.60169 -74.13956 594679.0 9049077 PIF_04668_2024
#> 4    11   Imiria -8.60166 -74.13892 594749.4 9049080 PIF_04668_2024
#> 5    11   Imiria -8.60117 -74.14487 594094.8 9049136 PIF_04668_2024
#> 6    11   Imiria -8.60107 -74.14316 594283.0 9049147 PIF_04668_2024
#>   created_user        created_date last_edited_user    last_edited_date ESTADO
#> 1   JMONTALBAN 2024-10-18 17:07:23       JMONTALBAN 2024-10-18 17:07:23      1
#> 2   JMONTALBAN 2024-10-18 17:07:23       JMONTALBAN 2024-10-18 17:07:23      1
#> 3   JMONTALBAN 2024-10-18 17:07:23       JMONTALBAN 2024-10-18 17:07:23      1
#> 4   JMONTALBAN 2024-10-18 17:07:23       JMONTALBAN 2024-10-18 17:07:23      1
#> 5   JMONTALBAN 2024-10-18 17:07:23       JMONTALBAN 2024-10-18 17:07:23      1
#> 6   JMONTALBAN 2024-10-18 17:07:23       JMONTALBAN 2024-10-18 17:07:23      1
#>               TIPCOB PELIGRO REFESTADO CONF CODRVIF SENSAT CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal       0         0    0       0    N20     25   2501 250104
#> 2 Cobertura Forestal       0         0    0       0    N21     25   2501 250104
#> 3 Cobertura Forestal       0         0    0       0    N20     25   2501 250104
#> 4 Cobertura Forestal       0         0    0       0    N21     25   2501 250104
#> 5 Cobertura Forestal       0         0    0       0    N21     25   2501 250104
#> 6 Cobertura Forestal       0         0    0       0    N20     25   2501 250104
#>                     geometry
#> 1 POINT (-74.14829 -8.60201)
#> 2 POINT (-74.14085 -8.60177)
#> 3 POINT (-74.13956 -8.60169)
#> 4 POINT (-74.13892 -8.60166)
#> 5 POINT (-74.14487 -8.60117)
#> 6 POINT (-74.14316 -8.60107)
# }