Download available forest fire data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_forest_fire_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
fire_forest <- get_forest_fire_data(show_progress = FALSE)
#> Reading layer `file1a9976eddc28' from data source 
#>   `/tmp/RtmpguWDyp/file1a9976eddc28.geojson' using driver `GeoJSON'
#> Simple feature collection with 448 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -78.77948 ymin: -16.37708 xmax: -69.20894 ymax: -3.28337
#> Geodetic CRS:  WGS 84
head(fire_forest)
#> Simple feature collection with 6 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -74.76679 ymin: -14.32693 xmax: -71.30557 ymax: -8.32145
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1   211588 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 2   211718 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 3   211836 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 4   211939 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 5   211589 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 6   211590 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN  NOMDEP           NOMPRO     NOMDIS      CAPITAL      FECHA
#> 1     19      2   CUSCO            CANAS     QUEHUE       QUEHUE 2024-10-08
#> 2     18      2 UCAYALI CORONEL PORTILLO    MASISEA      MASISEA 2024-10-08
#> 3     18      2 UCAYALI CORONEL PORTILLO   MANANTAY SAN FERNANDO 2024-10-08
#> 4     18      2 UCAYALI CORONEL PORTILLO CAMPOVERDE   CAMPOVERDE 2024-10-08
#> 5     19      2   CUSCO            CANAS     QUEHUE       QUEHUE 2024-10-08
#> 6     19      2   CUSCO          CANCHIS    SICUANI      SICUANI 2024-10-08
#>       HORA CATEG
#> 1 14:13:00    25
#> 2 13:51:00    14
#> 3 14:28:00    30
#> 4 13:51:00    30
#> 5 14:13:00    30
#> 6 13:49:00    30
#>                                                                            NOMCATEG
#> 1                                                                       CHAUPIBAMBA
#> 2 Consorcio Industrias G.L.C. San Juan E.I.R.L. - Empresa Forestal Shongfu E.I.R.L.
#> 3                                                                                  
#> 4                                                                                  
#> 5                                                                                  
#> 6                                                                                  
#>     LATITUD  LONGITUD   COORES  COORNO         CODREP created_user
#> 1 -14.32693 -71.55318 224593.8 8414597 PIF_04496_2024   JMONTALBAN
#> 2  -9.19567 -73.62495 651069.0 8983225 PIF_04515_2024   JMONTALBAN
#> 3  -8.44679 -74.57555 546722.1 9066283 PIF_04527_2024   JMONTALBAN
#> 4  -8.32145 -74.76679 525679.1 9080158 PIF_04529_2024   JMONTALBAN
#> 5 -14.32504 -71.54449 225529.4 8414817 PIF_04496_2024   JMONTALBAN
#> 6 -14.29447 -71.30557 251280.7 8418470 PIF_04497_2024   JMONTALBAN
#>          created_date last_edited_user    last_edited_date ESTADO
#> 1 2024-10-09 02:48:26           UFMS17 2024-10-09 03:51:27      2
#> 2 2024-10-09 02:48:26       JMONTALBAN 2024-10-09 02:48:26      1
#> 3 2024-10-09 02:48:26           UFMS17 2024-10-09 04:03:02      2
#> 4 2024-10-09 02:48:26           UFMS17 2024-10-09 04:47:34      1
#> 5 2024-10-09 02:48:26           UFMS17 2024-10-09 03:51:27      2
#> 6 2024-10-09 02:48:26           UFMS17 2024-10-09 03:52:36      2
#>               TIPCOB PELIGRO REFESTADO CONF CODRVIF SENSAT CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal       0         0    0       0      N     08   0805 080507
#> 2 Cobertura Forestal       0         0    0       0    N21     25   2501 250104
#> 3 Cobertura Forestal       0         0    0       0      A     25   2501 250107
#> 4 Cobertura Agrícola       0         0    0       0    N21     25   2501 250102
#> 5 Cobertura Forestal       0         0    0       0      N     08   0805 080507
#> 6 Cobertura Forestal       0         0    0       0    N21     08   0806 080601
#>                      geometry
#> 1 POINT (-71.55318 -14.32693)
#> 2  POINT (-73.62495 -9.19567)
#> 3  POINT (-74.57555 -8.44679)
#> 4  POINT (-74.76679 -8.32145)
#> 5 POINT (-71.54449 -14.32504)
#> 6 POINT (-71.30557 -14.29447)
# }