Download available forest fire data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_forest_fire_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
fire_forest <- get_forest_fire_data(show_progress = FALSE)
#> Reading layer `file1a4d7fc78265' from data source 
#>   `/tmp/RtmphDMLyY/file1a4d7fc78265.geojson' using driver `GeoJSON'
#> Simple feature collection with 5678 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.49888 ymin: -15.61481 xmax: -69.10229 ymax: -3.56854
#> Geodetic CRS:  WGS 84
head(fire_forest)
#> Simple feature collection with 6 features and 34 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -74.48153 ymin: -8.59718 xmax: -74.47054 ymax: -8.59503
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1   195126 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-17
#> 2   195127 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-17
#> 3   195128 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-17
#> 4   195129 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-17
#> 5   195130 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-17
#> 6   195131 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-17
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN  NOMDEP           NOMPRO   NOMDIS      CAPITAL      FECHA
#> 1     18      2 UCAYALI CORONEL PORTILLO MANANTAY SAN FERNANDO 2024-09-17
#> 2     18      2 UCAYALI CORONEL PORTILLO MANANTAY SAN FERNANDO 2024-09-17
#> 3     18      2 UCAYALI CORONEL PORTILLO MANANTAY SAN FERNANDO 2024-09-17
#> 4     18      2 UCAYALI CORONEL PORTILLO MANANTAY SAN FERNANDO 2024-09-17
#> 5     18      2 UCAYALI CORONEL PORTILLO MANANTAY SAN FERNANDO 2024-09-17
#> 6     18      2 UCAYALI CORONEL PORTILLO MANANTAY SAN FERNANDO 2024-09-17
#>       HORA CATEG      NOMCATEG  LATITUD  LONGITUD   COORES  COORNO
#> 1 14:10:00    24 NUEVO PARAISO -8.59718 -74.47883 557346.6 9049643
#> 2 13:43:00    24 NUEVO PARAISO -8.59653 -74.48153 557049.6 9049715
#> 3 14:10:00    24 NUEVO PARAISO -8.59642 -74.47469 557802.3 9049726
#> 4 13:43:00    24 NUEVO PARAISO -8.59578 -74.47691 557558.1 9049798
#> 5 14:10:00    24 NUEVO PARAISO -8.59565 -74.47054 558259.0 9049811
#> 6 13:43:00    24 NUEVO PARAISO -8.59503 -74.47227 558068.8 9049880
#>           CODREP created_user        created_date last_edited_user
#> 1 PIF_02967_2024       USUFMS 2024-09-18 16:00:13           UFMS17
#> 2 PIF_02967_2024       USUFMS 2024-09-18 16:00:13           UFMS17
#> 3 PIF_02967_2024       USUFMS 2024-09-18 16:00:13           UFMS17
#> 4 PIF_02967_2024       USUFMS 2024-09-18 16:00:13           UFMS17
#> 5 PIF_02967_2024       USUFMS 2024-09-18 16:00:13           UFMS17
#> 6 PIF_02967_2024       USUFMS 2024-09-18 16:00:13           UFMS17
#>      last_edited_date ESTADO             TIPCOB PELIGRO REFESTADO CONF CODRVIF
#> 1 2024-09-20 06:20:59      2 Cobertura Forestal       0         0    0       0
#> 2 2024-09-20 06:20:59      2 Cobertura Forestal       0         0    0       0
#> 3 2024-09-20 06:20:59      2 Cobertura Forestal       0         0    0       0
#> 4 2024-09-20 06:20:59      2 Cobertura Forestal       0         0    0       0
#> 5 2024-09-20 06:20:59      2 Cobertura Forestal       0         0    0       0
#> 6 2024-09-20 06:20:59      2 Cobertura Forestal       0         0    0       0
#>   SENSAT CATDEP CATPRO CATDIS                   geometry
#> 1      N     25   2501 250107 POINT (-74.47883 -8.59718)
#> 2    N21     25   2501 250107 POINT (-74.48153 -8.59653)
#> 3      N     25   2501 250107 POINT (-74.47469 -8.59642)
#> 4    N21     25   2501 250107 POINT (-74.47691 -8.59578)
#> 5      N     25   2501 250107 POINT (-74.47054 -8.59565)
#> 6    N21     25   2501 250107 POINT (-74.47227 -8.59503)
# }