Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1ab731ffad88' from data source 
#>   `/tmp/Rtmpxr6CpH/file1ab731ffad88.geojson' using driver `GeoJSON'
#> Simple feature collection with 3396 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.92632 ymin: -17.3824 xmax: -68.89607 ymax: -0.24211
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -79.07869 ymin: -6.926011 xmax: -78.25224 ymax: -5.003616
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1  9869981 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 2  9869982 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 3  9869983 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 4  9869984 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 5  9869985 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#> 6  9869986 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-18
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN    NOMDEP      NOMPRO       NOMDIS      CAPITAL      FECHA
#> 1     17      2  AMAZONAS   UTCUBAMBA      JAMALCA BAGUA GRANDE 2024-10-18
#> 2     17      2  AMAZONAS   UTCUBAMBA BAGUA GRANDE BAGUA GRANDE 2024-10-18
#> 3     17      2  AMAZONAS       BAGUA      LA PECA        BAGUA 2024-10-18
#> 4     17      2 CAJAMARCA   SAN PABLO     TUMBADEN    SAN PABLO 2024-10-18
#> 5     17      2 CAJAMARCA SAN IGNACIO  SAN IGNACIO  SAN IGNACIO 2024-10-18
#> 6     17      2 CAJAMARCA SAN IGNACIO     NAMBALLE  SAN IGNACIO 2024-10-18
#>       HORA CATEG NOMCATEG  LATITUD  LONGITUD   COORES  COORNO created_user
#> 1 10:20:00    30          -5.86600 -78.25224 804281.0 9350862       USUFMS
#> 2 11:00:00    30          -5.86600 -78.32411 796317.0 9350901       USUFMS
#> 3 12:50:00    25  LA PECA -5.57854 -78.46784 780530.6 9382782       USUFMS
#> 4 10:50:00    30          -6.92601 -78.62954 761934.2 9233777       USUFMS
#> 5 03:10:00    30          -5.05752 -78.97090 724961.8 9440627       USUFMS
#> 6 10:20:00    30          -5.00362 -79.07869 713023.5 9446624       USUFMS
#>          created_date last_edited_user    last_edited_date             TIPCOB
#> 1 2024-10-18 18:31:18           USUFMS 2024-10-18 18:31:18 Cobertura Agrícola
#> 2 2024-10-18 18:31:18           USUFMS 2024-10-18 18:31:18 Cobertura Agrícola
#> 3 2024-10-18 18:31:18           USUFMS 2024-10-18 18:31:18 Cobertura Forestal
#> 4 2024-10-18 18:31:18           USUFMS 2024-10-18 18:31:18 Cobertura Forestal
#> 5 2024-10-18 18:31:18           USUFMS 2024-10-18 18:31:18 Cobertura Agrícola
#> 6 2024-10-18 18:31:18           USUFMS 2024-10-18 18:31:18 Cobertura Forestal
#>           SENSAT PELIGRO CATDEP CATPRO CATDIS                    geometry
#> 1 GOES (16 y 18)       0     01   0107 010705 POINT (-78.25224 -5.865999)
#> 2 GOES (16 y 18)       0     01   0107 010701 POINT (-78.32411 -5.865999)
#> 3 GOES (16 y 18)       0     01   0102 010206 POINT (-78.46784 -5.578538)
#> 4 GOES (16 y 18)       0     06   0612 061204 POINT (-78.62954 -6.926011)
#> 5 GOES (16 y 18)       0     06   0609 060901  POINT (-78.9709 -5.057515)
#> 6 GOES (16 y 18)       0     06   0609 060905 POINT (-79.07869 -5.003616)
# }