Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1fa2ab1eaf0' from data source 
#>   `/tmp/RtmpwDDU1B/file1fa2ab1eaf0.geojson' using driver `GeoJSON'
#> Simple feature collection with 83 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.80846 ymin: -16.64268 xmax: -71.36332 ymax: -3.88629
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.06684 ymin: -15.46001 xmax: -74.51525 ymax: -6.15346
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 10992149 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 10992150 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 10992151 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 10992152 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 10992153 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 10992154 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-05 05:00:00
#> 2 2025-03-05 05:00:00
#> 3 2025-03-05 05:00:00
#> 4 2025-03-05 05:00:00
#> 5 2025-03-05 05:00:00
#> 6 2025-03-05 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN     NOMDEP     NOMPRO       NOMDIS    CAPITAL
#> 1     17      2 LAMBAYEQUE LAMBAYEQUE        OLMOS LAMBAYEQUE
#> 2     18      2   AREQUIPA   CARAVELI        ACARI   CARAVELI
#> 3     18      2   AREQUIPA   CARAVELI  BELLA UNION   CARAVELI
#> 4     18      2   AREQUIPA   CARAVELI  BELLA UNION   CARAVELI
#> 5     18      2        ICA      NASCA VISTA ALEGRE      NAZCA
#> 6     18      2        ICA      NASCA      MARCONA      NAZCA
#>                 FECHA     HORA CATEG NOMCATEG   LATITUD  LONGITUD   COORES
#> 1 2025-03-05 05:00:00 08:10:00    30           -6.15346 -80.06684 603247.4
#> 2 2025-03-05 05:00:00 13:10:00    30          -15.46001 -74.51525 552001.4
#> 3 2025-03-05 05:00:00 12:50:00    30          -15.17255 -74.71288 530842.5
#> 4 2025-03-05 05:00:00 12:50:00    30          -15.15458 -74.74881 526984.9
#> 5 2025-03-05 05:00:00 11:50:00    30          -14.97492 -74.98238 501894.8
#> 6 2025-03-05 05:00:00 11:20:00    30          -15.04678 -75.00034 499963.1
#>    COORNO created_user        created_date last_edited_user    last_edited_date
#> 1 9319741       USUFMS 2025-03-09 00:31:00           USUFMS 2025-03-09 00:31:00
#> 2 8290736       USUFMS 2025-03-09 00:31:00           USUFMS 2025-03-09 00:31:00
#> 3 8322569       USUFMS 2025-03-09 00:31:00           USUFMS 2025-03-09 00:31:00
#> 4 8324561       USUFMS 2025-03-09 00:31:00           USUFMS 2025-03-09 00:31:00
#> 5 8344448       USUFMS 2025-03-09 00:31:00           USUFMS 2025-03-09 00:31:00
#> 6 8336500       USUFMS 2025-03-09 00:31:00           USUFMS 2025-03-09 00:31:00
#>               TIPCOB         SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal GOES (16 y 18)       0     14   1403 140308
#> 2 Cobertura Forestal GOES (16 y 18)       0     04   0403 040302
#> 3 Cobertura Forestal GOES (16 y 18)       0     04   0403 040305
#> 4 Cobertura Forestal GOES (16 y 18)       0     04   0403 040305
#> 5 Cobertura Forestal GOES (16 y 18)       0     11   1103 110305
#> 6 Cobertura Forestal GOES (16 y 18)       0     11   1103 110304
#>                      geometry
#> 1  POINT (-80.06684 -6.15346)
#> 2 POINT (-74.51525 -15.46001)
#> 3 POINT (-74.71288 -15.17255)
#> 4 POINT (-74.74881 -15.15458)
#> 5 POINT (-74.98238 -14.97492)
#> 6 POINT (-75.00034 -15.04678)
# }