Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1fd2344a1afd' from data source 
#>   `/tmp/RtmpYpuDqB/file1fd2344a1afd.geojson' using driver `GeoJSON'
#> Simple feature collection with 144 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.83939 ymin: -16.74363 xmax: -69.2907 ymax: -3.88702
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.83939 ymin: -14.92102 xmax: -75.09017 ymax: -4.805987
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 10986605 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 10986606 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 10986607 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 10986608 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 10986609 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 10986610 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-04 05:00:00
#> 2 2025-03-04 05:00:00
#> 3 2025-03-04 05:00:00
#> 4 2025-03-04 05:00:00
#> 5 2025-03-04 05:00:00
#> 6 2025-03-04 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN NOMDEP  NOMPRO           NOMDIS  CAPITAL               FECHA
#> 1     17      2 ANCASH   SANTA           NEPEÑA CHIMBOTE 2025-03-04 05:00:00
#> 2     17      2  PIURA SULLANA IGNACIO ESCUDERO  SULLANA 2025-03-04 05:00:00
#> 3     18      2    ICA   NASCA            NASCA    NAZCA 2025-03-04 05:00:00
#> 4     18      2    ICA   NASCA            NASCA    NAZCA 2025-03-04 05:00:00
#> 5     18      2    ICA   NASCA            NASCA    NAZCA 2025-03-04 05:00:00
#> 6     18      2    ICA   NASCA            NASCA    NAZCA 2025-03-04 05:00:00
#>       HORA CATEG NOMCATEG   LATITUD  LONGITUD   COORES  COORNO created_user
#> 1 11:40:00    30           -9.18977 -78.30614 796040.8 8983056       USUFMS
#> 2 14:40:00    30           -4.80599 -80.83939 517809.2 9468779       USUFMS
#> 3 11:20:00    30          -14.84915 -75.09017 490298.8 8358356       USUFMS
#> 4 11:40:00    30          -14.92102 -75.10814 488369.8 8350407       USUFMS
#> 5 12:40:00    30          -14.79525 -75.12611 486429.7 8364316       USUFMS
#> 6 11:30:00    30          -14.86712 -75.12611 486434.2 8356367       USUFMS
#>          created_date last_edited_user    last_edited_date             TIPCOB
#> 1 2025-03-05 03:31:14           USUFMS 2025-03-05 03:31:14 Cobertura Agrícola
#> 2 2025-03-05 03:31:14           USUFMS 2025-03-05 03:31:14 Cobertura Forestal
#> 3 2025-03-05 03:31:14           USUFMS 2025-03-05 03:31:14 Cobertura Forestal
#> 4 2025-03-05 03:31:14           USUFMS 2025-03-05 03:31:14 Cobertura Forestal
#> 5 2025-03-05 03:31:14           USUFMS 2025-03-05 03:31:14 Cobertura Forestal
#> 6 2025-03-05 03:31:14           USUFMS 2025-03-05 03:31:14 Cobertura Forestal
#>           SENSAT PELIGRO CATDEP CATPRO CATDIS                    geometry
#> 1 GOES (16 y 18)       0     02   0218 021806 POINT (-78.30614 -9.189765)
#> 2 GOES (16 y 18)       0     20   2006 200603 POINT (-80.83939 -4.805987)
#> 3 GOES (16 y 18)       0     11   1103 110301 POINT (-75.09017 -14.84915)
#> 4 GOES (16 y 18)       0     11   1103 110301 POINT (-75.10814 -14.92102)
#> 5 GOES (16 y 18)       0     11   1103 110301 POINT (-75.12611 -14.79525)
#> 6 GOES (16 y 18)       0     11   1103 110301 POINT (-75.12611 -14.86712)
# }