Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1fbc5797dd96' from data source 
#>   `/tmp/RtmpV5J33r/file1fbc5797dd96.geojson' using driver `GeoJSON'
#> Simple feature collection with 61 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.23383 ymin: -18.23891 xmax: -70.50337 ymax: -3.9469
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -72.91293 ymin: -18.23891 xmax: -70.50337 ymax: -16.52123
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 10982725 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 10982726 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 10982727 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 10982728 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 10982729 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 10982730 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-02 05:00:00
#> 2 2025-03-02 05:00:00
#> 3 2025-03-02 05:00:00
#> 4 2025-03-02 05:00:00
#> 5 2025-03-02 05:00:00
#> 6 2025-03-02 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN   NOMDEP        NOMPRO              NOMDIS   CAPITAL
#> 1     19      2    TACNA         TACNA LA YARADA LOS PALOS LOS PALOS
#> 2     19      2    TACNA JORGE BASADRE                 ITE       ITE
#> 3     19      2    TACNA JORGE BASADRE             LOCUMBA   LOCUMBA
#> 4     19      2 AREQUIPA         ISLAY       DEAN VALDIVIA  LA CURVA
#> 5     18      2 AREQUIPA        CAMANA    MARISCAL CACERES  SAN JOSE
#> 6     18      2 AREQUIPA        CAMANA    MARISCAL CACERES  SAN JOSE
#>                 FECHA     HORA CATEG NOMCATEG   LATITUD  LONGITUD   COORES
#> 1 2025-03-02 05:00:00 12:38:00    30          -18.23891 -70.50337 341050.2
#> 2 2025-03-01 05:00:00 12:56:00    30          -17.87987 -71.04465 283364.3
#> 3 2025-03-02 05:00:00 12:38:00    30          -17.63954 -70.79686 309371.3
#> 4 2025-03-01 05:00:00 01:16:00    13     SN03 -17.13198 -71.87373 194219.1
#> 5 2025-03-02 05:00:00 13:55:00    30          -16.62422 -72.76654 738256.2
#> 6 2025-03-01 05:00:00 12:56:00    30          -16.52123 -72.91293 722752.4
#>    COORNO created_user        created_date last_edited_user    last_edited_date
#> 1 7982729       USUFMS 2025-03-03 06:06:39           USUFMS 2025-03-03 06:06:39
#> 2 8021918       USUFMS 2025-03-03 06:06:39           USUFMS 2025-03-03 06:06:39
#> 3 8048788       USUFMS 2025-03-03 06:06:39           USUFMS 2025-03-03 06:06:39
#> 4 8103584       USUFMS 2025-03-03 06:06:39           USUFMS 2025-03-03 06:06:39
#> 5 8160686       USUFMS 2025-03-03 06:06:39           USUFMS 2025-03-03 06:06:39
#> 6 8172254       USUFMS 2025-03-03 06:06:39           USUFMS 2025-03-03 06:06:39
#>               TIPCOB SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Agrícola    N20       0     23   2301 230111
#> 2 Cobertura Forestal    N20       0     23   2303 230303
#> 3 Cobertura Forestal    N20       0     23   2303 230301
#> 4 Cobertura Agrícola    N21       0     04   0407 040703
#> 5 Cobertura Agrícola      N       0     04   0402 040204
#> 6 Cobertura Forestal    N20       0     04   0402 040204
#>                      geometry
#> 1 POINT (-70.50337 -18.23891)
#> 2 POINT (-71.04465 -17.87987)
#> 3 POINT (-70.79686 -17.63954)
#> 4 POINT (-71.87373 -17.13198)
#> 5 POINT (-72.76654 -16.62422)
#> 6 POINT (-72.91293 -16.52123)
# }