Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1a8347c27d09' from data source 
#>   `/tmp/Rtmpwh3zuB/file1a8347c27d09.geojson' using driver `GeoJSON'
#> Simple feature collection with 5937 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.28617 ymin: -17.49401 xmax: -68.65796 ymax: -2.1015
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -70.69673 ymin: -17.49401 xmax: -69.20894 ymax: -16.37489
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1  9587745 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 2  9587746 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 3  9587747 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 4  9587748 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 5  9587749 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#> 6  9587750 Servicio Nacional Forestal y de Fauna Silvestre        2024-10-08
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN NOMDEP        NOMPRO  NOMDIS CAPITAL      FECHA     HORA CATEG
#> 1     19      2  TACNA JORGE BASADRE LOCUMBA LOCUMBA 2024-10-07 12:49:00    30
#> 2     19      2   PUNO      CHUCUITO  POMATA  POMATA 2024-10-08 12:56:00    25
#> 3     19      2   PUNO      CHUCUITO  POMATA  POMATA 2024-10-08 12:32:00    25
#> 4     19      2   PUNO      CHUCUITO  POMATA  POMATA 2024-10-08 13:49:00    25
#> 5     19      2   PUNO      CHUCUITO  POMATA  POMATA 2024-10-08 12:32:00     5
#> 6     19      2   PUNO      CHUCUITO  POMATA  POMATA 2024-10-08 13:49:00     5
#>                            NOMCATEG   LATITUD  LONGITUD   COORES  COORNO
#> 1                                   -17.49401 -70.69673 319852.7 8064992
#> 2                           TUQUINA -16.37756 -69.21733 476791.6 8189288
#> 3                          TAMBILLO -16.37708 -69.21427 477118.4 8189341
#> 4                           TUQUINA -16.37692 -69.21902 476611.1 8189358
#> 5 Reserva Paisajística Cerro Khapia -16.37656 -69.20942 477636.2 8189399
#> 6 Reserva Paisajística Cerro Khapia -16.37489 -69.20894 477687.3 8189584
#>   created_user        created_date last_edited_user    last_edited_date
#> 1       USUFMS 2024-10-09 01:17:50           USUFMS 2024-10-09 01:17:50
#> 2       USUFMS 2024-10-09 01:17:50           USUFMS 2024-10-09 01:17:50
#> 3       USUFMS 2024-10-09 01:17:50           USUFMS 2024-10-09 01:17:50
#> 4       USUFMS 2024-10-09 01:17:50           USUFMS 2024-10-09 01:17:50
#> 5       USUFMS 2024-10-09 01:17:50           USUFMS 2024-10-09 01:17:50
#> 6       USUFMS 2024-10-09 01:17:50           USUFMS 2024-10-09 01:17:50
#>               TIPCOB SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal      N       0     23   2303 230301
#> 2 Cobertura Agrícola    N20       0     21   2104 210406
#> 3 Cobertura Forestal      N       0     21   2104 210406
#> 4 Cobertura Agrícola    N21       0     21   2104 210406
#> 5 Cobertura Forestal      N       0     21   2104 210406
#> 6 Cobertura Forestal    N21       0     21   2104 210406
#>                      geometry
#> 1 POINT (-70.69673 -17.49401)
#> 2 POINT (-69.21733 -16.37756)
#> 3 POINT (-69.21427 -16.37708)
#> 4 POINT (-69.21902 -16.37692)
#> 5 POINT (-69.20942 -16.37656)
#> 6 POINT (-69.20894 -16.37489)
# }