Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1fab68241f87' from data source 
#>   `/tmp/RtmpJbiOJL/file1fab68241f87.geojson' using driver `GeoJSON'
#> Simple feature collection with 204 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.23383 ymin: -17.87987 xmax: -71.04465 ymax: -4.36884
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -79.29429 ymin: -8.201619 xmax: -78.917 ymax: -7.536865
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 10979202 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 10979203 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 10979204 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 10979205 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 10979206 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 10979207 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-01 05:00:00
#> 2 2025-03-01 05:00:00
#> 3 2025-03-01 05:00:00
#> 4 2025-03-01 05:00:00
#> 5 2025-03-01 05:00:00
#> 6 2025-03-01 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN      NOMDEP   NOMPRO      NOMDIS  CAPITAL               FECHA
#> 1     17      2 LA LIBERTAD TRUJILLO   SALAVERRY TRUJILLO 2025-03-01 05:00:00
#> 2     17      2 LA LIBERTAD   ASCOPE CASA GRANDE   ASCOPE 2025-03-01 05:00:00
#> 3     17      2 LA LIBERTAD   ASCOPE CASA GRANDE   ASCOPE 2025-03-01 05:00:00
#> 4     17      2 LA LIBERTAD   ASCOPE CASA GRANDE   ASCOPE 2025-03-01 05:00:00
#> 5     17      2 LA LIBERTAD   ASCOPE CASA GRANDE   ASCOPE 2025-03-01 05:00:00
#> 6     17      2 LA LIBERTAD   ASCOPE CASA GRANDE   ASCOPE 2025-03-01 05:00:00
#>       HORA CATEG NOMCATEG  LATITUD  LONGITUD   COORES  COORNO created_user
#> 1 13:30:00    30          -8.20162 -78.91700 729479.8 9092818       USUFMS
#> 2 13:10:00    25 TRINIDAD -7.57280 -79.25836 692150.8 9162545       USUFMS
#> 3 13:10:00    30          -7.59076 -79.25836 692142.9 9160558       USUFMS
#> 4 13:30:00    25 TRINIDAD -7.53687 -79.27632 690183.8 9166527       USUFMS
#> 5 13:10:00    25 TRINIDAD -7.55483 -79.25836 692158.8 9164532       USUFMS
#> 6 13:10:00    30          -7.60873 -79.29429 688169.7 9158586       USUFMS
#>          created_date last_edited_user    last_edited_date             TIPCOB
#> 1 2025-03-02 02:31:03           USUFMS 2025-03-02 02:31:03 Cobertura Forestal
#> 2 2025-03-02 02:31:03           USUFMS 2025-03-02 02:31:03 Cobertura Forestal
#> 3 2025-03-02 02:31:03           USUFMS 2025-03-02 02:31:03 Cobertura Forestal
#> 4 2025-03-02 02:31:03           USUFMS 2025-03-02 02:31:03 Cobertura Forestal
#> 5 2025-03-02 02:31:03           USUFMS 2025-03-02 02:31:03 Cobertura Forestal
#> 6 2025-03-02 02:31:03           USUFMS 2025-03-02 02:31:03 Cobertura Forestal
#>           SENSAT PELIGRO CATDEP CATPRO CATDIS                    geometry
#> 1 GOES (16 y 18)       0     13   1301 130109   POINT (-78.917 -8.201619)
#> 2 GOES (16 y 18)       0     13   1302 130208 POINT (-79.25836 -7.572798)
#> 3 GOES (16 y 18)       0     13   1302 130208 POINT (-79.25836 -7.590764)
#> 4 GOES (16 y 18)       0     13   1302 130208 POINT (-79.27632 -7.536865)
#> 5 GOES (16 y 18)       0     13   1302 130208 POINT (-79.25836 -7.554832)
#> 6 GOES (16 y 18)       0     13   1302 130208  POINT (-79.29429 -7.60873)
# }