Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1f9373ba482e' from data source 
#>   `/tmp/RtmpgZkd4Z/file1f9373ba482e.geojson' using driver `GeoJSON'
#> Simple feature collection with 24 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.62512 ymin: -16.37729 xmax: -69.2907 ymax: -3.88702
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -76.78957 ymin: -16.37729 xmax: -69.2907 ymax: -12.29057
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 10988228 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 10988229 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 10988230 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 10988231 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 10988232 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 10988233 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-05 05:00:00
#> 2 2025-03-05 05:00:00
#> 3 2025-03-05 05:00:00
#> 4 2025-03-05 05:00:00
#> 5 2025-03-05 05:00:00
#> 6 2025-03-05 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN   NOMDEP   NOMPRO        NOMDIS       CAPITAL
#> 1     19      2 AREQUIPA AREQUIPA          YURA          YURA
#> 2     19      2     PUNO     MOHO    HUAYRAPATA    HUAYRAPATA
#> 3     19      2     PUNO     MOHO    HUAYRAPATA    HUAYRAPATA
#> 4     18      2      ICA    NASCA    EL INGENIO    EL INGENIO
#> 5     18      2      ICA  CHINCHA     EL CARMEN     EL CARMEN
#> 6     18      2     LIMA     LIMA PUNTA HERMOSA PUNTA HERMOSA
#>                 FECHA     HORA CATEG   NOMCATEG   LATITUD  LONGITUD   COORES
#> 1 2025-03-05 05:00:00 00:50:00    30            -16.37729 -71.69355 212270.1
#> 2 2025-03-04 05:00:00 12:51:00    25 HUAYRAPATA -15.26045 -69.29070 468785.6
#> 3 2025-03-04 05:00:00 12:51:00    25 HUAYRAPATA -15.25996 -69.29368 468465.6
#> 4 2025-03-04 05:00:00 13:40:00    30            -14.67854 -75.11502 487616.2
#> 5 2025-03-04 05:00:00 13:42:00    30            -13.56473 -76.16202 374272.2
#> 6 2025-03-04 05:00:00 13:42:00    25     CUCUYA -12.29057 -76.78957 305371.2
#>    COORNO created_user        created_date last_edited_user    last_edited_date
#> 1 8187422       USUFMS 2025-03-05 18:05:45           USUFMS 2025-03-05 18:05:45
#> 2 8312846       USUFMS 2025-03-05 18:05:45           USUFMS 2025-03-05 18:05:45
#> 3 8312900       USUFMS 2025-03-05 18:05:45           USUFMS 2025-03-05 18:05:45
#> 4 8377225       USUFMS 2025-03-05 18:05:45           USUFMS 2025-03-05 18:05:45
#> 5 8500112       USUFMS 2025-03-05 18:05:45           USUFMS 2025-03-05 18:05:45
#> 6 8640668       USUFMS 2025-03-05 18:05:45           USUFMS 2025-03-05 18:05:45
#>               TIPCOB SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal    N20       0     04   0401 040128
#> 2 Cobertura Forestal    N21       0     21   2109 210903
#> 3 Cobertura Forestal    N21       0     21   2109 210903
#> 4 Cobertura Forestal    N20       0     11   1103 110303
#> 5 Cobertura Agrícola    N20       0     11   1102 110205
#> 6 Cobertura Forestal    N20       0     15   1501 150126
#>                      geometry
#> 1 POINT (-71.69355 -16.37729)
#> 2  POINT (-69.2907 -15.26045)
#> 3 POINT (-69.29368 -15.25996)
#> 4 POINT (-75.11502 -14.67854)
#> 5 POINT (-76.16202 -13.56473)
#> 6 POINT (-76.78957 -12.29057)
# }