Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1fd35fde45fb' from data source 
#>   `/tmp/Rtmpq42McZ/file1fd35fde45fb.geojson' using driver `GeoJSON'
#> Simple feature collection with 73 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.23231 ymin: -17.60513 xmax: -69.51988 ymax: -3.76058
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -72.79501 ymin: -17.60513 xmax: -70.28134 ymax: -14.38352
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 11028861 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 11028862 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 11028863 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 11028864 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 11028865 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 11028866 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-23 05:00:00
#> 2 2025-03-23 05:00:00
#> 3 2025-03-23 05:00:00
#> 4 2025-03-23 05:00:00
#> 5 2025-03-23 05:00:00
#> 6 2025-03-23 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN   NOMDEP         NOMPRO           NOMDIS  CAPITAL
#> 1     19      2 MOQUEGUA MARISCAL NIETO         MOQUEGUA MOQUEGUA
#> 2     18      2 AREQUIPA         CAMANA MARISCAL CACERES SAN JOSE
#> 3     18      2 AREQUIPA         CAMANA MARISCAL CACERES SAN JOSE
#> 4     18      2 AREQUIPA         CAMANA MARISCAL CACERES SAN JOSE
#> 5     18      2 AREQUIPA         CAMANA MARISCAL CACERES SAN JOSE
#> 6     19      2     PUNO       AZANGARO           POTONI   POTONI
#>                 FECHA     HORA CATEG NOMCATEG   LATITUD  LONGITUD   COORES
#> 1 2025-03-22 05:00:00 13:53:00    30          -17.60513 -70.99329 288485.3
#> 2 2025-03-22 05:00:00 14:19:00    30          -16.61737 -72.76006 738956.2
#> 3 2025-03-22 05:00:00 13:56:00    30          -16.61553 -72.75883 739089.8
#> 4 2025-03-22 05:00:00 13:56:00    30          -16.61543 -72.76215 738735.6
#> 5 2025-03-22 05:00:00 12:39:00    30          -16.59089 -72.79501 735258.5
#> 6 2025-03-22 05:00:00 12:39:00    30          -14.38352 -70.28134 361849.5
#>    COORNO created_user        created_date last_edited_user    last_edited_date
#> 1 8052387       USUFMS 2025-03-23 18:06:35           USUFMS 2025-03-23 18:06:35
#> 2 8161436       USUFMS 2025-03-23 18:06:35           USUFMS 2025-03-23 18:06:35
#> 3 8161639       USUFMS 2025-03-23 18:06:35           USUFMS 2025-03-23 18:06:35
#> 4 8161654       USUFMS 2025-03-23 18:06:35           USUFMS 2025-03-23 18:06:35
#> 5 8164409       USUFMS 2025-03-23 18:06:35           USUFMS 2025-03-23 18:06:35
#> 6 8409474       USUFMS 2025-03-23 18:06:35           USUFMS 2025-03-23 18:06:35
#>               TIPCOB SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal    N21       0     18   1801 180101
#> 2 Cobertura Agrícola      N       0     04   0402 040204
#> 3 Cobertura Agrícola    N21       0     04   0402 040204
#> 4 Cobertura Agrícola    N21       0     04   0402 040204
#> 5 Cobertura Agrícola      N       0     04   0402 040204
#> 6 Cobertura Forestal      N       0     21   2102 210209
#>                      geometry
#> 1 POINT (-70.99329 -17.60513)
#> 2 POINT (-72.76006 -16.61737)
#> 3 POINT (-72.75883 -16.61553)
#> 4 POINT (-72.76215 -16.61543)
#> 5 POINT (-72.79501 -16.59089)
#> 6 POINT (-70.28134 -14.38352)
# }