Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1fe478c12088' from data source 
#>   `/tmp/RtmpxQFrFA/file1fe478c12088.geojson' using driver `GeoJSON'
#> Simple feature collection with 277 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.7163 ymin: -16.74061 xmax: -69.10822 ymax: -3.03402
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -72.76051 ymin: -16.74061 xmax: -71.29153 ymax: -16.62221
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 11013191 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 11013192 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 11013193 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 11013194 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 11013195 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 11013196 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-19 05:00:00
#> 2 2025-03-19 05:00:00
#> 3 2025-03-19 05:00:00
#> 4 2025-03-19 05:00:00
#> 5 2025-03-19 05:00:00
#> 6 2025-03-19 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN   NOMDEP                NOMPRO           NOMDIS    CAPITAL
#> 1     19      2 MOQUEGUA GENERAL SANCHEZ CERRO       LA CAPILLA LA CAPILLA
#> 2     18      2 AREQUIPA                CAMANA           QUILCA     QUILCA
#> 3     18      2 AREQUIPA                CAMANA    SAMUEL PASTOR   LA PAMPA
#> 4     18      2 AREQUIPA                CAMANA    SAMUEL PASTOR   LA PAMPA
#> 5     18      2 AREQUIPA                CAMANA MARISCAL CACERES   SAN JOSE
#> 6     18      2 AREQUIPA                CAMANA MARISCAL CACERES   SAN JOSE
#>                 FECHA     HORA CATEG NOMCATEG   LATITUD  LONGITUD   COORES
#> 1 2025-03-19 05:00:00 13:36:00    30          -16.74061 -71.29153 255694.4
#> 2 2025-03-19 05:00:00 13:36:00    30          -16.70514 -72.20479 798091.9
#> 3 2025-03-19 05:00:00 13:58:00    30          -16.64081 -72.69307 746076.0
#> 4 2025-03-19 05:00:00 13:58:00    30          -16.64024 -72.69013 746390.5
#> 5 2025-03-19 05:00:00 13:36:00    30          -16.62437 -72.76051 738899.6
#> 6 2025-03-19 05:00:00 13:58:00    30          -16.62221 -72.73139 742010.0
#>    COORNO created_user        created_date last_edited_user    last_edited_date
#> 1 8147732       USUFMS 2025-03-20 03:08:38           USUFMS 2025-03-20 03:08:38
#> 2 8150972       USUFMS 2025-03-20 03:08:38           USUFMS 2025-03-20 03:08:38
#> 3 8158760       USUFMS 2025-03-20 03:08:38           USUFMS 2025-03-20 03:08:38
#> 4 8158820       USUFMS 2025-03-20 03:08:38           USUFMS 2025-03-20 03:08:38
#> 5 8160662       USUFMS 2025-03-20 03:08:38           USUFMS 2025-03-20 03:08:38
#> 6 8160866       USUFMS 2025-03-20 03:08:38           USUFMS 2025-03-20 03:08:38
#>               TIPCOB SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal      N       0     18   1802 180205
#> 2 Cobertura Forestal      N       0     04   0402 040207
#> 3 Cobertura Agrícola    N20       0     04   0402 040208
#> 4 Cobertura Agrícola    N20       0     04   0402 040208
#> 5 Cobertura Agrícola      N       0     04   0402 040204
#> 6 Cobertura Agrícola    N20       0     04   0402 040204
#>                      geometry
#> 1 POINT (-71.29153 -16.74061)
#> 2 POINT (-72.20479 -16.70514)
#> 3 POINT (-72.69307 -16.64081)
#> 4 POINT (-72.69013 -16.64024)
#> 5 POINT (-72.76051 -16.62437)
#> 6 POINT (-72.73139 -16.62221)
# }