Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: Serfor Platform

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1f80c610782' from data source 
#>   `/tmp/Rtmp3WuC5H/file1f80c610782.geojson' using driver `GeoJSON'
#> Simple feature collection with 144 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -80.83939 ymin: -16.74363 xmax: -69.2907 ymax: -3.88702
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -71.94124 ymin: -16.74363 xmax: -71.9277 ymax: -16.73856
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG
#> 1 10987160 Servicio Nacional Forestal y de Fauna Silvestre       
#> 2 10987161 Servicio Nacional Forestal y de Fauna Silvestre       
#> 3 10987162 Servicio Nacional Forestal y de Fauna Silvestre       
#> 4 10987163 Servicio Nacional Forestal y de Fauna Silvestre       
#> 5 10987164 Servicio Nacional Forestal y de Fauna Silvestre       
#> 6 10987165 Servicio Nacional Forestal y de Fauna Silvestre       
#>                FECREG
#> 1 2025-03-04 05:00:00
#> 2 2025-03-04 05:00:00
#> 3 2025-03-04 05:00:00
#> 4 2025-03-04 05:00:00
#> 5 2025-03-04 05:00:00
#> 6 2025-03-04 05:00:00
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN   NOMDEP NOMPRO   NOMDIS  CAPITAL               FECHA     HORA
#> 1     19      2 AREQUIPA  ISLAY MOLLENDO MOLLENDO 2025-03-03 05:00:00 13:36:00
#> 2     19      2 AREQUIPA  ISLAY MOLLENDO MOLLENDO 2025-03-03 05:00:00 13:36:00
#> 3     19      2 AREQUIPA  ISLAY MOLLENDO MOLLENDO 2025-03-03 05:00:00 13:36:00
#> 4     19      2 AREQUIPA  ISLAY MOLLENDO MOLLENDO 2025-03-03 05:00:00 13:36:00
#> 5     19      2 AREQUIPA  ISLAY MOLLENDO MOLLENDO 2025-03-03 05:00:00 13:36:00
#> 6     19      2 AREQUIPA  ISLAY MOLLENDO MOLLENDO 2025-03-03 05:00:00 13:36:00
#>   CATEG NOMCATEG   LATITUD  LONGITUD   COORES  COORNO created_user
#> 1    30          -16.74363 -71.93629 186914.9 8146493       USUFMS
#> 2    30          -16.74217 -71.92770 187829.1 8146668       USUFMS
#> 3    30          -16.74075 -71.94124 186382.0 8146804       USUFMS
#> 4    30          -16.74002 -71.93694 186839.6 8146892       USUFMS
#> 5    30          -16.73929 -71.93264 187297.3 8146979       USUFMS
#> 6    30          -16.73856 -71.92834 187754.9 8147067       USUFMS
#>          created_date last_edited_user    last_edited_date             TIPCOB
#> 1 2025-03-05 08:06:41           USUFMS 2025-03-05 08:06:41 Cobertura Forestal
#> 2 2025-03-05 08:06:41           USUFMS 2025-03-05 08:06:41 Cobertura Forestal
#> 3 2025-03-05 08:06:41           USUFMS 2025-03-05 08:06:41 Cobertura Forestal
#> 4 2025-03-05 08:06:41           USUFMS 2025-03-05 08:06:41 Cobertura Forestal
#> 5 2025-03-05 08:06:41           USUFMS 2025-03-05 08:06:41 Cobertura Forestal
#> 6 2025-03-05 08:06:41           USUFMS 2025-03-05 08:06:41 Cobertura Forestal
#>   SENSAT PELIGRO CATDEP CATPRO CATDIS                    geometry
#> 1      N       0     04   0407 040701 POINT (-71.93629 -16.74363)
#> 2      N       0     04   0407 040701  POINT (-71.9277 -16.74217)
#> 3      N       0     04   0407 040701 POINT (-71.94124 -16.74075)
#> 4      N       0     04   0407 040701 POINT (-71.93694 -16.74002)
#> 5      N       0     04   0407 040701 POINT (-71.93264 -16.73929)
#> 6      N       0     04   0407 040701 POINT (-71.92834 -16.73856)
# }