Download available hot spot data from Serfor's Satellite Monitoring Unit

This function allows you to download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru. For more information, please visit the following website: https://sniffs.serfor.gob.pe/monitoreo/sami/index.html

Usage

get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = FALSE)

Arguments

dsn

Character. Output filename with the spatial format. If missing, a temporary file is created.

show_progress

Logical. Suppress bar progress.

quiet

Logical. Suppress info message.

Value

An sf object.

Examples

# \donttest{
library(geoidep)
library(sf)
hot_spots <- get_hotspots_data(show_progress = FALSE)
#> Reading layer `file1a873a8d6b99' from data source 
#>   `/tmp/Rtmp2fmnJA/file1a873a8d6b99.geojson' using driver `GeoJSON'
#> Simple feature collection with 2373 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.2865 ymin: -17.47071 xmax: -68.76164 ymax: -0.7455
#> Geodetic CRS:  WGS 84
head(hot_spots)
#> Simple feature collection with 6 features and 29 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -78.57564 ymin: -8.309416 xmax: -73.9583 ymax: -3.799874
#> Geodetic CRS:  WGS 84
#>   OBJECTID                                          FUENTE DOCREG     FECREG
#> 1  7268175 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-18
#> 2  7268176 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-18
#> 3  7268177 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-18
#> 4  7268178 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-18
#> 5  7268179 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-18
#> 6  7268180 Servicio Nacional Forestal y de Fauna Silvestre        2024-09-18
#>                                                                                        OBSERV
#> 1 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 2 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 3 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 4 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 5 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#> 6 Información generada por la Unidad Funcional de Monitoreo Satelital de la DGIOFFS de SERFOR
#>   ZONUTM ORIGEN    NOMDEP           NOMPRO    NOMDIS              CAPITAL
#> 1     17      2  AMAZONAS     CONDORCANQUI EL CENEPA SANTA MARIA DE NIEVA
#> 2     17      2 CAJAMARCA        CAJAMARCA  ENCAÑADA            CAJAMARCA
#> 3     17      2 CAJAMARCA        CAJAMARCA  ENCAÑADA            CAJAMARCA
#> 4     17      2 CAJAMARCA        CAJAMARCA CAJAMARCA            CAJAMARCA
#> 5     17      2 CAJAMARCA        CAJAMARCA CAJAMARCA            CAJAMARCA
#> 6     18      2   UCAYALI CORONEL PORTILLO  CALLERIA             PUCALLPA
#>        FECHA     HORA CATEG        NOMCATEG  LATITUD  LONGITUD   COORES  COORNO
#> 1 2024-09-18 09:50:00    13            RC07 -3.79987 -78.34208 795220.2 9579540
#> 2 2024-09-18 10:40:00    30                 -6.92601 -78.55767 769879.8 9233736
#> 3 2024-09-18 10:20:00    30                 -6.90804 -78.57564 767903.5 9235734
#> 4 2024-09-18 10:50:00    30                 -6.92601 -78.57564 767893.4 9233747
#> 5 2024-09-18 10:20:00    30                 -6.94398 -78.57564 767883.2 9231759
#> 6 2024-09-18 09:30:00    14 K&B Company SAC -8.30942 -73.95830 614712.7 9081345
#>   created_user        created_date last_edited_user    last_edited_date
#> 1       USUFMS 2024-09-18 16:42:30           USUFMS 2024-09-18 16:42:30
#> 2       USUFMS 2024-09-18 16:42:30           USUFMS 2024-09-18 16:42:30
#> 3       USUFMS 2024-09-18 16:42:30           USUFMS 2024-09-18 16:42:30
#> 4       USUFMS 2024-09-18 16:42:30           USUFMS 2024-09-18 16:42:30
#> 5       USUFMS 2024-09-18 16:42:30           USUFMS 2024-09-18 16:42:30
#> 6       USUFMS 2024-09-18 16:42:30           USUFMS 2024-09-18 16:42:30
#>               TIPCOB         SENSAT PELIGRO CATDEP CATPRO CATDIS
#> 1 Cobertura Forestal GOES (16 y 17)       0     01   0104 010402
#> 2 Cobertura Forestal GOES (16 y 17)       0     06   0601 060105
#> 3 Cobertura Forestal GOES (16 y 17)       0     06   0601 060105
#> 4 Cobertura Forestal GOES (16 y 17)       0     06   0601 060101
#> 5 Cobertura Agrícola GOES (16 y 17)       0     06   0601 060101
#> 6 Cobertura Forestal GOES (16 y 17)       0     25   2501 250101
#>                      geometry
#> 1 POINT (-78.34208 -3.799874)
#> 2 POINT (-78.55767 -6.926011)
#> 3 POINT (-78.57564 -6.908045)
#> 4 POINT (-78.57564 -6.926011)
#> 5 POINT (-78.57564 -6.943977)
#> 6  POINT (-73.9583 -8.309416)
# }